Minor refactor on field aggregations
