+++ packages/graphql/src/translate/field-aggregations/utils.ts
@@ -1,4 +1,6 @@
 import { ResolveTree } from "graphql-parse-resolve-info";
+import { Node } from "../../classes";
+import { Context, RelationField } from "../../types";
 
 export enum AggregationType {
     Int = "IntAggregateSelection",
@@ -22,4 +24,12 @@ export function getFieldType(field: ResolveTree): AggregationType | undefined {
     if (field.fieldsByTypeName[AggregationType.Int]) return AggregationType.Int;
     if (field.fieldsByTypeName[AggregationType.String]) return AggregationType.String;
     return undefined;
+}
+
+export function wrapApocRun(query: string): string {
+    return `head(apoc.cypher.runFirstColumn(" ${query} ", {this:this}))`;
+}
+
+export function getReferenceNode(context: Context, relationField: RelationField): Node | undefined {
+    return context.neoSchema.nodes.find((x) => x.name === relationField.typeMeta.name);
 }
