+++ packages/graphql/tests/integration/aggregations/top-level/alias.int.test.ts
