+++ packages/graphql/tests/integration/aggregations/top-level/auth.int.test.ts
