+++ packages/graphql/tests/integration/aggregations/top-level/float.int.test.ts
