+++ packages/graphql/tests/integration/aggregations/top-level/string.int.test.ts
